+++ src/test/groovy/com/streamr/client/protocol/control_layer/UnicastMessageAdapterSpec.groovy
@@ -2,8 +2,8 @@ package com.streamr.client.protocol.control_layer
 
 import spock.lang.Specification
 
-import static com.streamr.client.protocol.StreamMessageExamples.InvalidSignature.helloWorld
-import static com.streamr.client.protocol.StreamMessageExamples.InvalidSignature.helloWorldSerialized32
+import static com.streamr.client.protocol.message_layer.StreamMessageExamples.InvalidSignature.helloWorld
+import static com.streamr.client.protocol.message_layer.StreamMessageExamples.InvalidSignature.helloWorldSerialized32
 
 class UnicastMessageAdapterSpec extends Specification {
