+++ src/test/groovy/com/streamr/client/protocol/message_layer/StreamMessageExamples.groovy
@@ -1,9 +1,5 @@
-package com.streamr.client.protocol
+package com.streamr.client.protocol.message_layer
 
-import com.streamr.client.protocol.message_layer.Json
-import com.streamr.client.protocol.message_layer.MessageID
-import com.streamr.client.protocol.message_layer.MessageRef
-import com.streamr.client.protocol.message_layer.StreamMessage
 import com.streamr.client.utils.Address
 
 class StreamMessageExamples {
